Big refactoring V: rework transfer and transit
Renamed filetransfer to transfer. Merged core::transit into transit. The API was redone and things got split up. Now the transit is completely independent of the transfer.
