+++ Cargo.toml
@@ -19,7 +19,7 @@ readme = "README.md"
 travis-ci = { repository = "warner/magic-wormhole.rs" }
 
 [dependencies]
-serde = "1.0"
+serde = { version = "1.0", features = ["rc"] }
 serde_json = "1.0"
 serde_derive = "1.0"
 xsalsa20poly1305 = "0.6"
